fix: Improve initialisms support
There were two issues I discovered with the initialism implementation:

1. Pluralized initialisms had no support, so while "ID" would be handled
   correctly, "IDs" would not. Because the "s" at the end of the
   initialism is traditionally lower-cased, adding a new config entry
   with "ids" would not have solved the problem.
2. Identifiers starting with initialisms incorrectly only had their
   first characters adjusted, when in practice their entire first word
   should have been set to lowercase. This resulted in some instances of
   `iD` seen in the test cases, instead of `id`.

This corrects both of those issues, and adds a fair chunk of unit tests
to prove that the logic is working as expected, considering the relative
complexity of each of those operations.

I think in a more ideal implementation, we wouldn't call `toLowerCase` on the
result of `StructName` but instead on the input to `StructName`. In
particular that could handle `xml_api`/`XMLAPI` correctly, where this
implementation doesn't. That said, the change as written seems like a
strict improvement over the current behavior without having to do a
larger refactor.